PF: Infer min. and max. sampling bounds
Set low and high parameters for random number generation automatically based on the minimum and maximum values in the input tensor. This helps to sample values in the same interval and avoid generating artificially high numbers which could perturb the results.
Initially values were sampled between 0.0 and 1.0 but this is not accurate because due to the mean and standard deviation of the dataset which is reduced from each image, the values end up in another interval different than between 0.0 and 1.0.