PF: Fix bug in perturbation mode random
In perturbation mode random, using the accumulative mask for flipping is not necessary
and causes a matplotlib warning regarding range of flipped input when all perturbation steps are executed.

The accumulative mask is only required for inpainting because inpainting results in similar values as in the original image.
Thus, the need to inpaint regions multiple times to remove as much from the original image as possible.